L1Location 的使用 MapId的型態 由 short -> int 簡化 FaceToFace 的 if判斷
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@1576 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/net/l1j/server/clientpackets/C_War.java
@@ -129,7 +129,7 @@ public class C_War extends ClientBasePacket {
 			int castle_id = enemyClan.getCastleId();
 			if (WarTimeController.getInstance().isNowWar(castle_id)) { // 戰爭時間內
 				L1PcInstance clanMember[] = clan.getOnlineClanMember();
-				for (short k = 0; k < clanMember.length; k++) {
+				for (int k = 0; k < clanMember.length; k++) {
 					if (L1CastleLocation.checkInWarArea(castle_id, clanMember[k])) {
 						player.sendPackets(new S_ServerMessage(SystemMessageId.$477));
 						return;
