+++ src/net/l1j/server/model/L1HauntedHouse.java
@@ -93,7 +93,7 @@ public class L1HauntedHouse {
 			if (pc.getMapId() == 5140) {
 				SkillUse skilluse = new SkillUse();
 				skilluse.handleCommands(pc, SKILL_CANCEL_MAGIC, pc.getId(), pc.getX(), pc.getY(), null, 0, Base.SKILL_TYPE[1]);
-				L1Teleport.teleport(pc, 32624, 32813, (short) 4, 5, true);
+				L1Teleport.teleport(pc, 32624, 32813, 4, 5, true);
 			}
 		}
 		clearMembers();
